+++ Examples/Apps/Treadmill/Software/Libs/Sources/SettingsSerializer.cpp
@@ -52,7 +52,7 @@ bool SettingsSerializer::save(const Settings &settings)
     const char *slash = strrchr(mPath, '/');
     if (slash) {
         char buff[SDK::Interface::IFileSystem::skMaxPathLen] { };
-        snprintf(buff, sizeof(buff), "%.*s", static_cast<size_t>(slash - mPath), mPath);
+        snprintf(buff, sizeof(buff), "%.*s", static_cast<int>(slash - mPath), mPath);
         // Create dir
         if (!mKernel.fs.mkdir(buff)) {
             return false;
